+++ src/components/Header.js
@@ -1,0 +1,12 @@
+import Row from 'react-bootstrap/Row';
+import Col from 'react-bootstrap/Col';
+
+const Header = () => (
+	<Row className='justify-content-md-center text-center'>
+		<Col>
+			<h1>Memory Game</h1>
+		</Col>
+	</Row>
+);
+
+export default Header;
