+++ src/helpers/shuffleCards.js
@@ -1,0 +1,8 @@
+export const shuffleCards = (cards) => {
+	const shuffledCards = cards.slice();
+	for (let i = shuffledCards.length - 1; i > 0; i--) {
+		const j = Math.floor(Math.random() * (i + 1));
+		[shuffledCards[i], shuffledCards[j]] = [shuffledCards[j], shuffledCards[i]];
+	}
+	return shuffledCards;
+};
