+++ src/helpers/storedUsername.js
@@ -1,0 +1,3 @@
+export const storedUsername = () => {
+	return localStorage.getItem('userName');
+};
